infrastructure layer refactored
